# public/api-doc/openapi.json
{
    "openapi": "3.0.0",
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "OAuth2"
            }
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "",
            "variables": {}
        }
    ],
    "info": {
        "version": "",
        "title": "Dự Án Tốt Nghiệp API document",
        "description": "",
        "termsOfService": "",
        "contact": {}
    },
    "paths": {
...
        "/api/post/categories/{id}": {
            "get": {
                "tags": ["Post (Bài viết)"],
                "summary": "Get list Post by Category ID",
                "operationId": "get-list-post-by-category-id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Category id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 4,
                                                "post_title": "Vespa GTS phiên bản hoàn toàn mới ra mắt: thay đổi nhẹ ở thiết kế, đồng hồ LCD, thêm Smartkey",
                                                "post_slug": "post-slug",
                                                "post_desc": "Tập đoàn Piaggio vừa giới thiệu dòng sản phẩm Vespa GTS thế hệ hoàn toàn mới, với tuỳ chọn động cơ trải dài từ 125 cho tới 300. Có thể nói đây là đợt nâng cấp lớn đối với Vespa GTS bởi có khá nhiều chi tiết trên xe được cải tiến và cũng thay đổi luôn thói quen sử dụng xe. Tóm lược về những sản phẩm thuộc dòng GTS thì chúng ta sẽ có 4 nhánh lớn, trong đó bao gồm GTS, GTS Super, GTS SuperSport và GTS SuperTech. Phân loại này cũng tương tự như phiên bản GTS hiện có. Cả 4 dòng GTS đều sẽ có 2 tuỳ chọn động cơ, đó là động cơ iGet 125 và động cơ HPE 300. Tại Ý, giá bán của các dòng xe cụ thể như sau:",
                                                "date": null,
                                                "category_name": "Tin công nghệ"
                                            },
                                            {
                                                "id": 2,
                                                "post_title": "Xiaomi ra mắt loạt smartphone, tai nghe, smartwatch, TV và thiết bị làm sạch",
                                                "post_slug": "Xiaomi",
                                                "post_desc": "Xiaomi 12T và 12T Pro là bộ đôi smartphone cao cấp vừa được Xiaomi ra mắt nổi bật với hệ thống camera có độ phân giải cao cùng hiệu năng mạnh mẽ, cạnh tranh với nhiều đối thủ trong cùng phân khúc giá.",
                                                "date": "2022-10-12T16:24:28.000000Z",
                                                "category_name": "Tin công nghệ"
                                            }
                                        ],
                                        "first_page_url": "http://127.0.0.1:8000/api/post/categories/1?page=1",
                                        "from": 1,
                                        "last_page": 1,
                                        "last_page_url": "http://127.0.0.1:8000/api/post/categories/1?page=1",
                                        "links": [
                                            {
                                                "url": null,
                                                "label": "pagination.previous",
                                                "active": false
                                            },
                                            {
                                                "url": "http://127.0.0.1:8000/api/post/categories/1?page=1",
                                                "label": "1",
                                                "active": true
                                            },
                                            {
                                                "url": null,
                                                "label": "pagination.next",
                                                "active": false
                                            }
                                        ],
                                        "next_page_url": null,
                                        "path": "http://127.0.0.1:8000/api/post/categories/1",
                                        "per_page": 20,
                                        "prev_page_url": null,
                                        "to": 2,
                                        "total": 2
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/product/{id}": {
            "get": {
                "tags": ["Product"],
                "summary": "Get Product By Id",
                "operationId": "get-product-by-id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Product id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": {
                                        "id": 2,
                                        "product_title": "Samsungx100 ",
                                        "product_desc": "Description",
                                        "product_content": "Content",
                                        "product_slug": "samsungx100",
                                        "product_image": "img.img",
                                        "product_quantity": 1,
                                        "product_price": 5000000,
                                        "product_promotion_desc": null,
                                        "product_images": [
                                            {
                                                "id": 1,
                                                "image": "img.img",
                                                "product_id": 2,
                                                "created_at": null,
                                                "updated_at": null,
                                                "deleted_at": null
                                            },
                                            {
                                                "id": 2,
                                                "image": "img2.img",
                                                "product_id": 2,
                                                "created_at": null,
                                                "updated_at": null,
                                                "deleted_at": null
                                            }
                                        ],
                                        "product_configurations": {
                                            "id": 1,
                                            "product_id": 2,
                                            "config_screen": "4.5",
                                            "config_cpu": "Android",
                                            "config_ram": 64,
                                            "config_camera": 120,
                                            "config_selfie": 12,
                                            "config_battery": 3800,
                                            "config_system": null,
                                            "created_at": null,
                                            "updated_at": null,
                                            "deleted_at": null
                                        },
                                        "is_variation": 0,
                                        "is_discount_product": 0,
                                        "properties": [
                                            {
                                                "id": 1,
                                                "propertie_name": "Màu sắc",
                                                "propertie_slug": "mau-sac",
                                                "propertie_value": "Xanh",
                                                "propertie_img": "img.img",
                                                "product_id": 2,
                                                "created_at": null,
                                                "updated_at": null,
                                                "deleted_at": null
                                            },
                                            {
                                                "id": 2,
                                                "propertie_name": "Dung lượng",
                                                "propertie_slug": "dung-luong",
                                                "propertie_value": "128 GB",
                                                "propertie_img": "img2.img",
                                                "product_id": 2,
                                                "created_at": null,
                                                "updated_at": null,
                                                "deleted_at": null
                                            },
                                            {
                                                "id": 3,
                                                "propertie_name": "Dung lượng",
                                                "propertie_slug": "dung-luong",
                                                "propertie_value": "64 GB",
                                                "propertie_img": "img.img",
                                                "product_id": 2,
                                                "created_at": null,
                                                "updated_at": null,
                                                "deleted_at": null
                                            }
                                        ],
                                        "category_name": "samsung2",
                                        "category_parent_id": 5,
                                        "product_suppiler_name": "Samsung Thái Nguyên",
                                        "product_seo_title": null,
                                        "product_seo_keyword": null,
                                        "product_seo_description": null
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/categories/{id}": {
            "get": {
                "tags": ["Product"],
                "summary": "Get List Product By Id Category",
                "operationId": "get-list-product-by-id-category",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Category id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "product_title": "Samsung J7",
                                                "product_slug": "samsung-j7",
                                                "product_desc": null,
                                                "product_price": 200000,
                                                "product_image": null,
                                                "product_configurations": null,
                                                "is_variation": 0,
                                                "is_discount_product": 0,
                                                "properties": [],
                                                "category_name": "Điện thoại",
                                                "category_parent_id": 0,
                                                "product_suppiler_name": "Samsung Thái Nguyên"
                                            }
                                        ],
                                        "first_page_url": "http://127.0.0.1:8000/api/product/categories/3?page=1",
                                        "from": 1,
                                        "last_page": 1,
                                        "last_page_url": "http://127.0.0.1:8000/api/product/categories/3?page=1",
                                        "links": [
                                            {
                                                "url": null,
                                                "label": "pagination.previous",
                                                "active": false
                                            },
                                            {
                                                "url": "http://127.0.0.1:8000/api/product/categories/3?page=1",
                                                "label": "1",
                                                "active": true
                                            },
                                            {
                                                "url": null,
                                                "label": "pagination.next",
                                                "active": false
                                            }
                                        ],
                                        "next_page_url": null,
                                        "path": "http://127.0.0.1:8000/api/product/categories/3",
                                        "per_page": 12,
                                        "prev_page_url": null,
                                        "to": 1,
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}